norm + the line(overdraw)

+++ ft_draw_minimap.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   ft_draw_minimap.c                                  :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: idabligi <idabligi@student.42.fr>          +#+  +:+       +#+        */
+/*   By: sbadr <sbadr@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/07/15 10:08:25 by idabligi          #+#    #+#             */
-/*   Updated: 2023/07/22 08:27:08 by idabligi         ###   ########.fr       */
+/*   Updated: 2023/07/24 23:35:21 by sbadr            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -16,8 +16,8 @@
 
 void	ft_draw_mm_line(t_data *data, double ongle)
 {
-	float	x;
-	float	y;
+	double	x;
+	double	y;
 	int		count;
 
 	count = 15;
